build: extract `git-branchless-revset` crate
