+++ git-branchless-opts/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 edition = "2021"
 name = "git-branchless-opts"
-version = "0.6.0-rc.1"
+version = "0.7.0-pre.1"
 
 [dependencies]
 clap = { version = "4.0.23", features = ["derive"] }
