+++ git-branchless-revset/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+edition = "2021"
+name = "git-branchless-revset"
+version = "0.7.0-pre.1"
+
+[dependencies]
+bstr = "1.0.1"
+chrono = "0.4.22"
+chrono-english = "0.1.7"
+chronoutil = "0.2.3"
+eden_dag = { package = "esl01-dag", version = "0.2.1" }
+eyre = "0.6.8"
+git-branchless-opts = { version = "0.7.0-pre.1", path = "../git-branchless-opts" }
+glob = "0.3.0"
+itertools = "0.10.5"
+lalrpop = "0.19.8"
+lalrpop-util = "0.19.8"
+lazy_static = "1.4.0"
+lib = { package = "git-branchless-lib", version = "0.7.0-pre.1", path = "../git-branchless-lib" }
+rayon = "1.5.3"
+regex = "1.6.0"
+thiserror = "1.0.32"
+tracing = "0.1.37"
+
+[build-dependencies]
+lalrpop = "0.19.8"
+
+[dev-dependencies]
+insta = "1.21.0"
