+++ git-branchless/Cargo.toml
@@ -35,7 +35,8 @@ dialoguer = "0.10.1"
 eden_dag = { package = "esl01-dag", version = "0.2.1" }
 eyre = "0.6.8"
 fslock = "0.2.1"
-git-branchless-opts = { version = "0.6.0-rc.1", path = "../git-branchless-opts" }
+git-branchless-opts = { version = "0.7.0-pre.1", path = "../git-branchless-opts" }
+git-branchless-revset = { version = "0.7.0-pre.1", path = "../git-branchless-revset" }
 git-record = { version = "0.3", path = "../git-record" }
 glob = "0.3.0"
 itertools = "0.10.3"
