+++ git-branchless/src/commands/bug_report.rs
@@ -12,7 +12,7 @@ use lib::core::repo_ext::{RepoExt, RepoReferencesSnapshot};
 use lib::util::ExitCode;
 
 use crate::commands::smartlog::{make_smartlog_graph, render_graph};
-use crate::revset::resolve_default_smartlog_commits;
+use git_branchless_revset::resolve_default_smartlog_commits;
 use lib::core::dag::Dag;
 use lib::core::effects::Effects;
 use lib::core::eventlog::{Event, EventCursor, EventLogDb, EventReplayer};
