+++ git-branchless/src/commands/restack.rs
@@ -66,8 +66,8 @@ use rayon::{ThreadPool, ThreadPoolBuilder};
 use tracing::{instrument, warn};
 
 use crate::commands::smartlog::smartlog;
-use crate::revset::resolve_commits;
 use git_branchless_opts::{MoveOptions, ResolveRevsetOptions, Revset};
+use git_branchless_revset::resolve_commits;
 use lib::core::config::get_restack_preserve_timestamps;
 use lib::core::dag::{commit_set_to_vec, union_all, CommitSet, Dag};
 use lib::core::effects::Effects;
