+++ git-branchless/src/commands/undo.rs
@@ -21,8 +21,8 @@ use tracing::instrument;
 
 use crate::commands::smartlog::{make_smartlog_graph, render_graph};
 use crate::declare_views;
-use crate::revset::resolve_default_smartlog_commits;
 use crate::tui::{with_siv, SingletonView};
+use git_branchless_revset::resolve_default_smartlog_commits;
 use lib::core::dag::{CommitSet, Dag};
 use lib::core::effects::Effects;
 use lib::core::eventlog::{Event, EventCursor, EventLogDb, EventReplayer, EventTransactionId};
